Refactor Vim plugin
